backend/pyproject.toml
[tool.poetry]
name = "linkedin-insight-hub"
version = "0.1.0"
description = "A project to visualize LinkedIn saved posts using embeddings and topic modeling"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "linkedin_insighthub" }]

[tool.poetry.dependencies]
python = "^3.11"

llvmlite = ">=0.39.1"
numba = ">=0.57.2"
numpy = ">=1.27"
pandas = ">=2.1"
scikit-learn = ">=1.3"
umap-learn = ">=0.5.4"
sentence-transformers = ">=2.2.2"
bertopic = ">=0.15.0"
wordcloud = "^1.9.5"
nltk = "^3.9.2"
fastapi = "^0.128.0"
vyper-config = "^1.2.1"
dotenv = "^0.9.9"
uvicorn = "^0.40.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7"
isort = "^6.0"
mypy = "^1.6"
pytest = "^7.4"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
